+++ tests/integration/staking/keeper/validator_test.go
@@ -376,7 +376,10 @@ func TestGetValidatorsEdgeCases(t *testing.T) {
 	// validator 3 enters bonded validator set
 	f.sdkCtx = f.sdkCtx.WithBlockHeight(40)
 
-	validators[3], err = f.stakingKeeper.GetValidator(f.sdkCtx, validators[3].GetOperator())
+	valbz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[3].GetOperator())
+	assert.NilError(t, err)
+
+	validators[3], err = f.stakingKeeper.GetValidator(f.sdkCtx, valbz)
 	assert.NilError(t, err)
 	f.stakingKeeper.DeleteValidatorByPowerIndex(f.sdkCtx, validators[3])
 	validators[3], _ = validators[3].AddTokensFromDel(f.stakingKeeper.TokensFromConsensusPower(f.sdkCtx, 1))
@@ -423,7 +426,7 @@ func TestGetValidatorsEdgeCases(t *testing.T) {
 	assert.Equal(t, nMax, uint32(len(resValidators)))
 	assert.Assert(ValEq(t, validators[0], resValidators[0]))
 	assert.Assert(ValEq(t, validators[2], resValidators[1]))
-	_, exists := f.stakingKeeper.GetValidator(f.sdkCtx, validators[3].GetOperator())
+	_, exists := f.stakingKeeper.GetValidator(f.sdkCtx, valbz)
 	assert.Assert(t, exists)
 }
 
@@ -495,7 +498,10 @@ func TestFullValidatorSetPowerChange(t *testing.T) {
 		keeper.TestingUpdateValidator(f.stakingKeeper, f.sdkCtx, validators[i], true)
 	}
 	for i := range powers {
-		validators[i], err = f.stakingKeeper.GetValidator(f.sdkCtx, validators[i].GetOperator())
+		valbz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[i].GetOperator())
+		assert.NilError(t, err)
+
+		validators[i], err = f.stakingKeeper.GetValidator(f.sdkCtx, valbz)
 		assert.NilError(t, err)
 	}
 	assert.Equal(t, types.Unbonded, validators[0].Status)
@@ -544,8 +550,12 @@ func TestApplyAndReturnValidatorSetUpdatesAllNone(t *testing.T) {
 	f.stakingKeeper.SetValidatorByPowerIndex(f.sdkCtx, validators[1])
 
 	updates := applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, 2)
-	validators[0], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[0].GetOperator())
-	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[1].GetOperator())
+	val0bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[0].GetOperator())
+	assert.NilError(t, err)
+	val1bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[1].GetOperator())
+	assert.NilError(t, err)
+	validators[0], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val0bz)
+	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val1bz)
 	assert.DeepEqual(t, validators[0].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[1])
 	assert.DeepEqual(t, validators[1].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 }
@@ -635,7 +645,9 @@ func TestApplyAndReturnValidatorSetUpdatesInserted(t *testing.T) {
 	f.stakingKeeper.SetValidator(f.sdkCtx, validators[2])
 	f.stakingKeeper.SetValidatorByPowerIndex(f.sdkCtx, validators[2])
 	updates := applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, 1)
-	validators[2], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[2].GetOperator())
+	val2bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[2].GetOperator())
+	assert.NilError(t, err)
+	validators[2], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val2bz)
 	assert.DeepEqual(t, validators[2].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 
 	// test validtor added at the beginning
@@ -643,7 +655,9 @@ func TestApplyAndReturnValidatorSetUpdatesInserted(t *testing.T) {
 	f.stakingKeeper.SetValidator(f.sdkCtx, validators[3])
 	f.stakingKeeper.SetValidatorByPowerIndex(f.sdkCtx, validators[3])
 	updates = applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, 1)
-	validators[3], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[3].GetOperator())
+	val3bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[3].GetOperator())
+	assert.NilError(t, err)
+	validators[3], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val3bz)
 	assert.DeepEqual(t, validators[3].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 
 	// test validtor added at the end
@@ -651,7 +665,9 @@ func TestApplyAndReturnValidatorSetUpdatesInserted(t *testing.T) {
 	f.stakingKeeper.SetValidator(f.sdkCtx, validators[4])
 	f.stakingKeeper.SetValidatorByPowerIndex(f.sdkCtx, validators[4])
 	updates = applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, 1)
-	validators[4], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[4].GetOperator())
+	val4bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[4].GetOperator())
+	assert.NilError(t, err)
+	validators[4], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val4bz)
 	assert.DeepEqual(t, validators[4].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 }
 
@@ -686,7 +702,9 @@ func TestApplyAndReturnValidatorSetUpdatesWithCliffValidator(t *testing.T) {
 	f.stakingKeeper.SetValidator(f.sdkCtx, validators[2])
 	f.stakingKeeper.SetValidatorByPowerIndex(f.sdkCtx, validators[2])
 	updates := applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, 2)
-	validators[2], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[2].GetOperator())
+	val2bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[2].GetOperator())
+	assert.NilError(t, err)
+	validators[2], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val2bz)
 	assert.DeepEqual(t, validators[0].ABCIValidatorUpdateZero(), updates[1])
 	assert.DeepEqual(t, validators[2].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 }
@@ -717,8 +735,13 @@ func TestApplyAndReturnValidatorSetUpdatesNewValidator(t *testing.T) {
 
 	// verify initial CometBFT updates are correct
 	updates := applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, len(validators))
-	validators[0], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[0].GetOperator())
-	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[1].GetOperator())
+
+	val0bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[0].GetOperator())
+	assert.NilError(t, err)
+	val1bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[1].GetOperator())
+	assert.NilError(t, err)
+	validators[0], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val0bz)
+	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val1bz)
 	assert.DeepEqual(t, validators[0].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 	assert.DeepEqual(t, validators[1].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[1])
 
@@ -762,9 +785,11 @@ func TestApplyAndReturnValidatorSetUpdatesNewValidator(t *testing.T) {
 
 	// verify initial CometBFT updates are correct
 	updates = applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, len(validators)+1)
-	validator, _ = f.stakingKeeper.GetValidator(f.sdkCtx, validator.GetOperator())
-	validators[0], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[0].GetOperator())
-	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[1].GetOperator())
+	valbz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validator.GetOperator())
+	assert.NilError(t, err)
+	validator, _ = f.stakingKeeper.GetValidator(f.sdkCtx, valbz)
+	validators[0], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val0bz)
+	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val1bz)
 	assert.DeepEqual(t, validator.ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 	assert.DeepEqual(t, validators[0].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[1])
 	assert.DeepEqual(t, validators[1].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[2])
@@ -796,8 +821,12 @@ func TestApplyAndReturnValidatorSetUpdatesBondTransition(t *testing.T) {
 
 	// verify initial CometBFT updates are correct
 	updates := applyValidatorSetUpdates(t, f.sdkCtx, f.stakingKeeper, 2)
-	validators[2], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[2].GetOperator())
-	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, validators[1].GetOperator())
+	val1bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[1].GetOperator())
+	assert.NilError(t, err)
+	val2bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[2].GetOperator())
+	assert.NilError(t, err)
+	validators[2], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val2bz)
+	validators[1], _ = f.stakingKeeper.GetValidator(f.sdkCtx, val1bz)
 	assert.DeepEqual(t, validators[2].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[0])
 	assert.DeepEqual(t, validators[1].ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx)), updates[1])
 
@@ -806,7 +835,9 @@ func TestApplyAndReturnValidatorSetUpdatesBondTransition(t *testing.T) {
 	// delegate to validator with lowest power but not enough to bond
 	f.sdkCtx = f.sdkCtx.WithBlockHeight(1)
 
-	validators[0], err = f.stakingKeeper.GetValidator(f.sdkCtx, validators[0].GetOperator())
+	val0bz, err := f.stakingKeeper.ValidatorAddressCodec().StringToBytes(validators[0].GetOperator())
+	assert.NilError(t, err)
+	validators[0], err = f.stakingKeeper.GetValidator(f.sdkCtx, val0bz)
 	assert.NilError(t, err)
 
 	f.stakingKeeper.DeleteValidatorByPowerIndex(f.sdkCtx, validators[0])
@@ -822,7 +853,7 @@ func TestApplyAndReturnValidatorSetUpdatesBondTransition(t *testing.T) {
 	// lowest power in a single block context (height)
 	f.sdkCtx = f.sdkCtx.WithBlockHeight(2)
 
-	validators[1], err = f.stakingKeeper.GetValidator(f.sdkCtx, validators[1].GetOperator())
+	validators[1], err = f.stakingKeeper.GetValidator(f.sdkCtx, val1bz)
 	assert.NilError(t, err)
 
 	f.stakingKeeper.DeleteValidatorByPowerIndex(f.sdkCtx, validators[0])